LayoutTests/http/tests/contentextensions/hide-on-csp-report-expected.txt

CONSOLE MESSAGE: Refused to load the image 'http://localhost/foo.png' because it violates the following Content Security Policy directive: "img-src 'self'".

This test creates a CSP violation report, but the report URL matches a 'css-display-none' rule.
This text should remain visible.

 

--------
Frame: 'result_frame'
--------
Ping received.
No cookies in ping.
